@@ -9,16 +9,13 @@ import android.view.ViewGroup;
 import android.view.Window;
 import android.widget.TextView;
 
-import com.bopr.android.smailer.R;
-import com.bopr.android.smailer.Settings;
-
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.FragmentActivity;
 
-import static com.bopr.android.smailer.util.TagFormatter.formatter;
+import com.bopr.android.smailer.BuildInfo;
+import com.bopr.android.smailer.R;
 
 /**
  * About dialog fragment.
@@ -27,42 +24,26 @@ import static com.bopr.android.smailer.util.TagFormatter.formatter;
  */
 public class AboutDialogFragment extends DialogFragment {
 
-    private Settings settings;
-
-    @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        settings = new Settings(requireContext());
-    }
-
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.dialog_about, container, false);
 
+        BuildInfo info = new BuildInfo(requireContext());
+
         TextView versionLabel = view.findViewById(R.id.label_message);
-        versionLabel.setText(formatVersion());
-
-        versionLabel.setOnLongClickListener(new View.OnLongClickListener() {
-
-            @Override
-            public boolean onLongClick(View v) {
-                Settings.BuildInfo info = settings.getReleaseInfo();
-                new AlertDialog.Builder(requireContext())
-                        .setTitle("Release info")
-                        .setMessage("Build number: " + info.number + "\nBuild time: " + info.time)
-                        .show();
-                return true;
-            }
+        versionLabel.setText(getString(R.string.app_version, info.name));
+        versionLabel.setOnLongClickListener(v -> {
+            new AlertDialog.Builder(requireContext())
+                    .setTitle("Release info")
+                    .setMessage("Build number: " + info.number + "\nBuild time: " + info.time)
+                    .show();
+            return true;
         });
 
-        view.findViewById(R.id.label_open_source).setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(getContext(), LegalInfoActivity.class));
-                dismiss();
-            }
+        view.findViewById(R.id.label_open_source).setOnClickListener(v -> {
+            startActivity(new Intent(getContext(), LegalInfoActivity.class));
+            dismiss();
         });
 
         return view;
@@ -74,12 +55,6 @@ public class AboutDialogFragment extends DialogFragment {
         Dialog dialog = super.onCreateDialog(savedInstanceState);
         dialog.requestWindowFeature(Window.FEATURE_NO_TITLE);
         return dialog;
-    }
-
-    private String formatVersion() {
-        return formatter(requireContext()).pattern(R.string.app_version)
-                .put("version", settings.getReleaseVersion())
-                .format();
     }
 
     static void showAboutDialog(@NonNull FragmentActivity activity) {
