@@ -4,7 +4,6 @@ import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.preference.Preference;
-import androidx.preference.Preference.OnPreferenceChangeListener;
 
 import com.bopr.android.smailer.R;
 import com.bopr.android.smailer.util.TagFormatter.TagPattern;
@@ -30,93 +29,57 @@ public class RulesFragment extends BasePreferenceFragment {
     public void onCreatePreferences(Bundle bundle, String rootKey) {
         addPreferencesFromResource(R.xml.pref_rules);
 
-        requirePreference(PREF_EMAIL_TRIGGERS).setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
-
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object value) {
-                if (value == null || ((Set) value).isEmpty()) {
-                    updateSummary(preference, getString(R.string.no_triggers_specified), SUMMARY_STYLE_ACCENTED);
-                } else {
-                    updateSummary(preference, getString(R.string.events_causing_sending_mail), SUMMARY_STYLE_DEFAULT);
-                }
-                return true;
+        requirePreference(PREF_EMAIL_TRIGGERS).setOnPreferenceChangeListener((preference, value) -> {
+            if (value == null || ((Set) value).isEmpty()) {
+                updateSummary(preference, getString(R.string.no_triggers_specified), SUMMARY_STYLE_ACCENTED);
+            } else {
+                updateSummary(preference, getString(R.string.events_causing_sending_mail), SUMMARY_STYLE_DEFAULT);
             }
+            return true;
         });
 
         Preference phoneBlacklistPreference = requirePreference(PREF_FILTER_PHONE_BLACKLIST);
-        phoneBlacklistPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                startActivity(new Intent(getContext(), PhoneBlacklistActivity.class));
-                return true;
-            }
+        phoneBlacklistPreference.setOnPreferenceClickListener(preference -> {
+            startActivity(new Intent(getContext(), PhoneBlacklistActivity.class));
+            return true;
         });
-        phoneBlacklistPreference.setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
-
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object value) {
-                updateSummary(preference, formatListAndSize((String) value, R.string.unacceptable_phone_numbers,
-                        R.string._none), SUMMARY_STYLE_DEFAULT);
-                return true;
-            }
+        phoneBlacklistPreference.setOnPreferenceChangeListener((preference, value) -> {
+            updateSummary(preference, formatListAndSize((String) value, R.string.unacceptable_phone_numbers,
+                    R.string._none), SUMMARY_STYLE_DEFAULT);
+            return true;
         });
 
         Preference phoneWhitelistPreference = requirePreference(PREF_FILTER_PHONE_WHITELIST);
-        phoneWhitelistPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                startActivity(new Intent(getContext(), PhoneWhitelistActivity.class));
-                return true;
-            }
+        phoneWhitelistPreference.setOnPreferenceClickListener(preference -> {
+            startActivity(new Intent(getContext(), PhoneWhitelistActivity.class));
+            return true;
         });
-        phoneWhitelistPreference.setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
-
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object value) {
-                updateSummary(preference, formatListAndSize((String) value, R.string.acceptable_phone_numbers,
-                        R.string._any), SUMMARY_STYLE_DEFAULT);
-                return true;
-            }
+        phoneWhitelistPreference.setOnPreferenceChangeListener((preference, value) -> {
+            updateSummary(preference, formatListAndSize((String) value, R.string.acceptable_phone_numbers,
+                    R.string._any), SUMMARY_STYLE_DEFAULT);
+            return true;
         });
 
         Preference textBlacklistPreference = requirePreference(PREF_FILTER_TEXT_BLACKLIST);
-        textBlacklistPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                startActivity(new Intent(getContext(), TextBlacklistActivity.class));
-                return true;
-            }
+        textBlacklistPreference.setOnPreferenceClickListener(preference -> {
+            startActivity(new Intent(getContext(), TextBlacklistActivity.class));
+            return true;
         });
-        textBlacklistPreference.setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
-
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object value) {
-                updateSummary(preference, formatListAndSize((String) value, R.string.unacceptable_words,
-                        R.string._none), SUMMARY_STYLE_DEFAULT);
-                return true;
-            }
+        textBlacklistPreference.setOnPreferenceChangeListener((preference, value) -> {
+            updateSummary(preference, formatListAndSize((String) value, R.string.unacceptable_words,
+                    R.string._none), SUMMARY_STYLE_DEFAULT);
+            return true;
         });
 
         Preference textWhitelistPreference = requirePreference(PREF_FILTER_TEXT_WHITELIST);
-        textWhitelistPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                startActivity(new Intent(getContext(), TextWhitelistActivity.class));
-                return true;
-            }
+        textWhitelistPreference.setOnPreferenceClickListener(preference -> {
+            startActivity(new Intent(getContext(), TextWhitelistActivity.class));
+            return true;
         });
-        textWhitelistPreference.setOnPreferenceChangeListener(new OnPreferenceChangeListener() {
-
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object value) {
-                updateSummary(preference, formatListAndSize((String) value, R.string.acceptable_words,
-                        R.string._any), SUMMARY_STYLE_DEFAULT);
-                return true;
-            }
+        textWhitelistPreference.setOnPreferenceChangeListener((preference, value) -> {
+            updateSummary(preference, formatListAndSize((String) value, R.string.acceptable_words,
+                    R.string._any), SUMMARY_STYLE_DEFAULT);
+            return true;
         });
 
     }
